1.09  + add skt default settings for all http server bindings  + add new option: fKeepAlive  + add new option: keepAliveIDLEtime  + add new option: keepAliveProbesInterval  + add new option: keepAliveNumberOfProbes  * http server lib upgraded to 1.6  * fix: http server: 'Content-Size' -> 'Content-Length'  * fix: linux: apply system limits before create thread pool  * fix: FreeBSD: thread pool may lost messages  * fAcceptFilter http server param does not work  * mmalloc_fd cant allocate >50% mem on FreeBSD without increace vm.max_wired  - remove option: fAcceptFilter for http server - allways ON

+++ src/include/old/core_rtimer.h
@@ -52,13 +52,13 @@ typedef struct rtq_timer_s {
 	uintptr_t	interval;	/* timer interval, 0 - if one shot timer (no repeats) */
 	rtq_cb		cb_func;	/* function that called at time. */
 	void		*udata;		/* user defined data, per thread pool. */
-	uintptr_t	flags;		/* RT_TIMER_F_* */
+	uint32_t	flags;		/* RT_TIMER_F_* */
 } rtq_timer_t;
 
 TAILQ_HEAD(rtq_timer_head, rtq_timer_s);
 
-#define RT_TIMER_F_ALLOCATED	(1 << 0) /* Internal use. */
-#define RT_TIMER_F_PERIODIC	(1 << 1)
+#define RT_TIMER_F_ALLOCATED	(((uint32_t)1) << 0) /* Internal use. */
+#define RT_TIMER_F_PERIODIC	(((uint32_t)1) << 1)
